% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\docType{package}
\name{tempconverter-package}
\alias{tempconverter}
\alias{tempconverter-package}
\title{tempconverter: Converting Fahrenheit to Celsius}
\description{
This package include information and functions on how to convert Fahrenheit to Celsius.
}
\author{
\strong{Maintainer}: Athena Flores Miranda \email{florea3@mcmaster.ca}

}
\keyword{internal}
